+++ tests/unit/contracts/test_layers.py
@@ -2,7 +2,7 @@ import pytest
 from grimp.adaptors.graph import ImportGraph
 
 from importlinter.application.app_config import settings
-from importlinter.contracts.layers import LayersContract
+from importlinter.contracts.layers import Layer, LayerField, LayersContract
 from importlinter.domain.contract import ContractCheck, InvalidContractOptions
 from importlinter.domain.helpers import MissingImport
 from tests.adapters.printing import FakePrinter
@@ -12,6 +12,22 @@ from tests.adapters.timing import FakeTimer
 @pytest.fixture(scope="module", autouse=True)
 def configure():
     settings.configure(TIMER=FakeTimer())
+
+
+@pytest.mark.parametrize(
+    "data, parsed_data",
+    (
+        ("one", Layer(name="one", is_optional=False)),
+        ("(one)", Layer(name="one", is_optional=True)),
+        ("one | two |    three", {Layer(name="one"), Layer(name="two"), Layer(name="three")}),
+        (
+            "one | (two) | three",
+            {Layer(name="one"), Layer(name="two", is_optional=True), Layer(name="three")},
+        ),
+    ),
+)
+def test_layer_field(data, parsed_data):
+    assert LayerField().parse(data) == parsed_data
 
 
 class TestLayerContractSingleContainers:
@@ -76,6 +92,65 @@ class TestLayerContractSingleContainers:
             session_options={"root_packages": ["mypackage"]},
             contract_options={"containers": ["mypackage"], "layers": ["high", "medium", "low"]},
         )
+
+
+class TestLayerContractSiblingLayers:
+    def test_illegal_imports(self):
+        contract = LayersContract(
+            name="Layer contract",
+            session_options={"root_packages": ["mypackage"]},
+            contract_options={
+                "containers": ["mypackage"],
+                "layers": ["high", "medium_a | medium_b | medium_c", "low"],
+            },
+        )
+        graph = ImportGraph()
+        for module in (
+            "mypackage",
+            "mypackage.high",
+            "mypackage.medium_a",
+            "mypackage.medium_b",
+            "mypackage.medium_c",
+            "mypackage.low",
+        ):
+            graph.add_module(module)
+        # Add some 'legal' imports.
+        graph.add_import(importer="mypackage.high.green", imported="mypackage.medium.orange")
+        graph.add_import(importer="mypackage.utils", imported="mypackage.medium.red")
+        # Add an illegal import.
+        graph.add_import(
+            importer="mypackage.medium_a.blue",
+            imported="mypackage.medium_b.red",
+            line_number=3,
+            line_contents="-",
+        )
+
+        contract_check = contract.check(graph=graph, verbose=False)
+
+        assert contract_check.kept is False
+        sorted_metadata = _get_sorted_metadata(contract_check)
+        assert sorted_metadata == {
+            "invalid_dependencies": [
+                {
+                    "importer": "mypackage.medium_a",
+                    "imported": "mypackage.medium_b",
+                    "routes": [
+                        {
+                            "chain": [
+                                {
+                                    "importer": "mypackage.medium_a.blue",
+                                    "imported": "mypackage.medium_b.red",
+                                    "line_numbers": (3,),
+                                },
+                            ],
+                            "extra_firsts": [],
+                            "extra_lasts": [],
+                        }
+                    ],
+                },
+            ],
+            "undeclared_modules": set(),
+        }
 
 
 class TestLayerMultipleContainers:
@@ -264,19 +339,13 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
         assert contract_check.kept is False
 
-        sorted_metadata = {
-            "invalid_chains": sorted(
-                contract_check.metadata["invalid_chains"],
-                key=lambda c: (c["lower_layer"], c["higher_layer"]),
-            ),
-            "undeclared_modules": contract_check.metadata["undeclared_modules"],
-        }
+        sorted_metadata = _get_sorted_metadata(contract_check)
         assert sorted_metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -301,9 +370,9 @@ class TestLayerContractPopulatesMetadata:
                     ],
                 },
                 {
-                    "higher_layer": "mypackage.medium",
-                    "lower_layer": "mypackage.low",
-                    "chains": [
+                    "imported": "mypackage.medium",
+                    "importer": "mypackage.low",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -323,9 +392,9 @@ class TestLayerContractPopulatesMetadata:
                     ],
                 },
                 {
-                    "lower_layer": "mypackage.medium",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.medium",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -365,11 +434,11 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -431,11 +500,11 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -491,11 +560,11 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -557,11 +626,11 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -630,11 +699,11 @@ class TestLayerContractPopulatesMetadata:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -724,11 +793,11 @@ class TestLayerContractPopulatesMetadata:
         assert contract_check.kept is False
 
         assert contract_check.metadata == {
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -821,11 +890,11 @@ class TestIgnoreImports:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.kept is False
-        assert contract_check.metadata["invalid_chains"] == [
+        assert contract_check.metadata["invalid_dependencies"] == [
             {
-                "lower_layer": "mypackage.low",
-                "higher_layer": "mypackage.medium",
-                "chains": [
+                "importer": "mypackage.low",
+                "imported": "mypackage.medium",
+                "routes": [
                     {
                         "chain": [
                             dict(
@@ -853,11 +922,11 @@ class TestIgnoreImports:
         contract_check = contract.check(graph=graph, verbose=False)
 
         assert contract_check.kept is False
-        assert contract_check.metadata["invalid_chains"] == [
+        assert contract_check.metadata["invalid_dependencies"] == [
             {
-                "higher_layer": "mypackage.medium",
-                "lower_layer": "mypackage.low",
-                "chains": [
+                "imported": "mypackage.medium",
+                "importer": "mypackage.low",
+                "routes": [
                     {
                         "chain": [
                             dict(
@@ -1074,11 +1143,11 @@ def test_render_broken_contract():
     check = ContractCheck(
         kept=False,
         metadata={
-            "invalid_chains": [
+            "invalid_dependencies": [
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -1168,9 +1237,9 @@ def test_render_broken_contract():
                     ],
                 },
                 {
-                    "lower_layer": "mypackage.low",
-                    "higher_layer": "mypackage.medium",
-                    "chains": [
+                    "importer": "mypackage.low",
+                    "imported": "mypackage.medium",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -1185,9 +1254,9 @@ def test_render_broken_contract():
                     ],
                 },
                 {
-                    "lower_layer": "mypackage.medium",
-                    "higher_layer": "mypackage.high",
-                    "chains": [
+                    "importer": "mypackage.medium",
+                    "imported": "mypackage.high",
+                    "routes": [
                         {
                             "chain": [
                                 {
@@ -1519,7 +1588,7 @@ class TestExhaustiveContracts:
 
         assert contract_check.kept is False
         assert contract_check.metadata == {
-            "invalid_chains": [],
+            "invalid_dependencies": [],
             "undeclared_modules": {"foo.blue"},
         }
 
@@ -1552,7 +1621,7 @@ class TestExhaustiveContracts:
 
         assert contract_check.kept is False
         assert contract_check.metadata == {
-            "invalid_chains": [],
+            "invalid_dependencies": [],
             "undeclared_modules": {"foo.blue", "foo.green"},
         }
 
@@ -1573,7 +1642,7 @@ class TestExhaustiveContracts:
 
         assert contract_check.kept is False
         assert contract_check.metadata == {
-            "invalid_chains": [],
+            "invalid_dependencies": [],
             "undeclared_modules": {"foo.green"},
         }
 
@@ -1633,7 +1702,7 @@ class TestExhaustiveContracts:
 
         assert contract_check.kept is False
         assert contract_check.metadata == {
-            "invalid_chains": [],
+            "invalid_dependencies": [],
             "undeclared_modules": {"bar.blue", "bar.yellow", "foo.blue", "foo.brown"},
         }
 
@@ -1649,3 +1718,13 @@ class TestExhaustiveContracts:
                 graph.add_module(module)
 
         return graph
+
+
+def _get_sorted_metadata(contract_check: ContractCheck) -> dict:
+    return {
+        "invalid_dependencies": sorted(
+            contract_check.metadata["invalid_dependencies"],
+            key=lambda c: (c["importer"], c["imported"]),
+        ),
+        "undeclared_modules": contract_check.metadata["undeclared_modules"],
+    }